Overlay Code: Rework reference counting and hook freeing
* We had rather complex logic to try to hide ourselves from the outside
world. This transparency comes with complexity in our code which is hard to
follow and thus error-prone. The enforced transparency is achieved through
lying.
* As a result, the reference counting is more consistent to the outside world
as well. The DX SDK Sample SimpleSample for example has refcount debugging
enabled and no longer indicates non-freed objects on exit.

* Remove the custom refCount logic and use the native IUnknown refcount logic
instead.
* Be aware of our own refCount influences (namely StateBlock pSB and Texture
texTexture) and clean up our objects when we are the only one left owning
references to the device.
** The clean up now includes device hooks.
* Clean up the hook freeing. It was incorrect
(writing on no-longer-existant object in memory).
* Add missing releaseStateBlock in createCleanState method
* Fix findOriginalDevice which was not releasing everything

* Add asserts
* Adjust DevState to make use of constructor and destructor
* Clean up DevState release methods
* In createDevice create own objects only once we know we need them